+++ testdata/apps/custom_main/app/setup.go
@@ -4,3 +4,5 @@ import "github.com/emersonjoe/trilha"
 
 func Setup(a *trilha.App) error    { return nil }
 func Shutdown(a *trilha.App) error { return nil }
+
+func Config(cfg *trilha.Config) error { return nil }